Update README.md
These visualizations can aid in the identification of key patterns and trends that may contribute to the occurrence of police killings in America.

This analysis is of great significance, given the current social and political climate surrounding issues of police brutality and systemic racism. Through a deeper understanding of the factors that influence police killings, we can develop more effective strategies to prevent them and promote greater accountability within law enforcement agencies.